# openapi/openapi-componentcatalog-v1.0.0.yaml
openapi: 3.0.3
info:
  title: Component Catalog REST API
  version: '1.0.0'
  description: >
    The Component Catalog API allows clients to retrieve information about CatalogItems, CatalogFilters and Files entities.
    
    Catalog and File entities also exist internally, but only referenced by id's on requests and not returned on responses.
    
    **NOTES**:
    - The OpenAPI specification file is also used to [generate](https://openapi-generator.tech/) REST client(s) and a server REST API.
    - Clients and servers generated from the same OpenAPI specification version are guaranteed to be **compatible**.
  contact:
    name: EDPCore Team
    url: https://confluence.biscrum.com/pages/viewpage.action?spaceKey=EDP&title=Welcome
servers:
  - url: http://{baseurl}/v1
    variables:
      baseurl:
        default: localhost:8080
        description: Default address for a Component Catalog's backend REST API instance.
security:
  - bearerAuth: []
...
paths:
  /catalog-descriptors:
    get:
      tags:
        - Catalog-descriptors
      summary: List of all available Catalog Descriptors.
      description: >
        Returns a list of all available Catalog Descriptors.<br/>
      operationId: getCatalogDescriptors
      responses:
        "200":
          description: A list of Catalog Descriptors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogDescriptor'
        "401":
          description: Invalid client token on the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "403":
          description: Insufficient permissions for the client to access the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestErrorMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: >
        Authorization via bearer token.
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CatalogDescriptor:
      properties:
        id:
          type: string
          example: 'aSdFam...yCg=='
        slug:
          type: string
          example: 'aSdFam...yCg=='
...
    RestErrorMessage:
      properties:
        message:
          type: string
      required:
        - message
...