# src/main/java/com/boehringer/componentcatalog/server/api/CatalogDescriptorsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.boehringer.componentcatalog.server.api;

import com.boehringer.componentcatalog.server.model.CatalogDescriptor;
import com.boehringer.componentcatalog.server.model.RestErrorMessage;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Controller
@Tag(name = "Catalog-descriptors", description = "the Catalog-descriptors API")
public interface CatalogDescriptorsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /catalog-descriptors : List of all available Catalog Descriptors.
     * Returns a list of all available Catalog Descriptors.&lt;br/&gt; 
     *
     * @return A list of Catalog Descriptors. (status code 200)
     *         or Invalid client token on the request. (status code 401)
     *         or Insufficient permissions for the client to access the resource. (status code 403)
     *         or Server error. (status code 500)
     */
    @Operation(
        operationId = "getCatalogDescriptors",
        summary = "List of all available Catalog Descriptors.",
        description = "Returns a list of all available Catalog Descriptors.<br/> ",
        tags = { "Catalog-descriptors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of Catalog Descriptors.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CatalogDescriptor.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid client token on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "403", description = "Insufficient permissions for the client to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/catalog-descriptors",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CatalogDescriptor>> getCatalogDescriptors(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : \"aSdFam...yCg==\", \"slug\" : \"aSdFam...yCg==\" }, { \"id\" : \"aSdFam...yCg==\", \"slug\" : \"aSdFam...yCg==\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}